Clean up Conference component
